project-codes/auth/requirements.txt
Django
djangorestframework
markdown
Django-filter
pillow
Django-guardian
coreapi
twilio

Django-cors-headers
